move Errors.kt from jvm to common
