+++ ktor-http/ktor-http-cio/api/ktor-http-cio.api
@@ -142,10 +142,6 @@ public final class io/ktor/http/cio/Response : io/ktor/http/cio/HttpMessage {
 	public final fun getVersion ()Ljava/lang/CharSequence;
 }
 
-public final class io/ktor/http/cio/UnsupportedMediaTypeExceptionCIO : java/io/IOException {
-	public fun <init> (Ljava/lang/String;)V
-}
-
 public final class io/ktor/http/cio/internals/CharsKt {
 	public static final fun parseDecLong (Ljava/lang/CharSequence;)J
 }
@@ -157,5 +153,9 @@ public final class io/ktor/http/cio/internals/MutableRange {
 	public final fun setEnd (I)V
 	public final fun setStart (I)V
 	public fun toString ()Ljava/lang/String;
+}
+
+public final class io/ktor/http/cio/internals/UnsupportedMediaTypeExceptionCIO : java/io/IOException {
+	public fun <init> (Ljava/lang/String;)V
 }
