+++ ktor-http/ktor-http-cio/common/src/io/ktor/http/cio/internals/Errors.kt
@@ -2,7 +2,7 @@
  * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.http.cio
+package io.ktor.http.cio.internals
 
 import io.ktor.utils.io.*
 import kotlinx.io.IOException
