+++ ktor-server/ktor-server-core/jvm/src/io/ktor/server/engine/DefaultTransformJvm.kt
@@ -6,6 +6,7 @@ package io.ktor.server.engine
 
 import io.ktor.http.*
 import io.ktor.http.cio.*
+import io.ktor.http.cio.internals.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.UnsupportedMediaTypeException
